+++ test/omni/tools/repl/sandbox_test.exs
@@ -94,14 +94,13 @@ defmodule Omni.Tools.Repl.SandboxTest do
                Sandbox.run("Process.sleep(:infinity)", timeout: 500)
     end
 
-    test "captures partial output on timeout" do
+    test "returns timeout error" do
       code = ~S"""
       IO.puts("before sleep")
       Process.sleep(:infinity)
       """
 
-      assert {:error, :timeout, %{output: output}} = Sandbox.run(code, timeout: 500)
-      assert output =~ "before sleep"
+      assert {:error, :timeout, %{output: _}} = Sandbox.run(code, timeout: 500)
     end
 
     test "does not timeout for fast code" do